type long for story.id, adding, comments

+++ src/main/java/mvc/mc/dh/model/Story.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class Story {
     @Getter
-    private final int ID;
+    private final long ID;
     @Getter
     private final String TITLE;
     @Getter
